Много полезного не помню уже с чего начал

+++ Kursovaya work/change_service.cs
@@ -25,53 +25,61 @@ namespace Kursovaya_work
         }
         private void button1_Click(object sender, EventArgs e)
         {
-            conbaza.Open();
+            if (textBox1.Text == "" || textBox2.Text == "" || textBox3.Text == "")
             {
-                string commandStr = $"INSERT INTO Service (ID_service,service, cost_service) VALUES (@id,@servic,@cost)";
-                MySqlCommand command = new MySqlCommand(commandStr, conbaza);
-                try
+                MessageBox.Show("Введите данные");
+            }
+            else
+            { 
+                conbaza.Open();
                 {
-                    //берём значение из текстбокса и кидаем в базу данных
-                    command.Parameters.Add("@id", MySqlDbType.VarChar).Value = textBox3.Text;
-                    command.Parameters.Add("@servic", MySqlDbType.VarChar).Value = textBox1.Text;
-                    command.Parameters.Add("@cost", MySqlDbType.VarChar).Value = textBox2.Text;                  
-                    //Изменения данных в базе данных
-                    command.ExecuteNonQuery();
-                    
-                }
-                catch (Exception ex)
-                {
-                    MessageBox.Show($"{ex}");
-                }
-                finally
-                {
-                    conbaza.Close();
+                    string commandStr = $"INSERT INTO Service (ID_service,service, cost_service) VALUES (@id,@servic,@cost)";
+                    MySqlCommand command = new MySqlCommand(commandStr, conbaza);
+                    try
+                    {
+                        //берём значение из текстбокса и кидаем в базу данных
+                        command.Parameters.Add("@id", MySqlDbType.VarChar).Value = textBox3.Text;
+                        command.Parameters.Add("@servic", MySqlDbType.VarChar).Value = textBox1.Text;
+                        command.Parameters.Add("@cost", MySqlDbType.VarChar).Value = textBox2.Text;
+                        //Изменения данных в базе данных
+                        command.ExecuteNonQuery();
+
+                    }
+                    catch (Exception ex)
+                    {
+                        MessageBox.Show($"{ex}");
+                    }
+                    finally
+                    {
+                        conbaza.Close();
+                    }
                 }
             }
         }
         
         private void button5_Click(object sender, EventArgs e)
-        {
-            try
-            {
-                conbaza.Open();
-                //Запрос к базе данных(взять ID_service,service,cose_service из таблицы service)
-                string commandStr = "SELECT ID_service,service,cost_service FROM Service";
-                MySqlDataAdapter adapter = new MySqlDataAdapter(commandStr, conbaza);
-                DataTable dTable = new DataTable();
-                adapter.Fill(dTable);
-                dataGridView1.Rows.Clear();
-                //Добавление строк пока i не станет больше или равно количеству строк таблицы(dTable.Rows.Count количество строк таблицы)
-                for (int i = 0; i < dTable.Rows.Count; i++)
+        {                 
+                try
                 {
-                    dataGridView1.Rows.Add(dTable.Rows[i].ItemArray);
+                    conbaza.Open();
+                    //Запрос к базе данных(взять ID_service,service,cose_service из таблицы service)
+                    string commandStr = "SELECT ID_service,service,cost_service FROM Service";
+                    MySqlDataAdapter adapter = new MySqlDataAdapter(commandStr, conbaza);
+                    DataTable dTable = new DataTable();
+                    adapter.Fill(dTable);
+                    dataGridView1.Rows.Clear();
+                    //Добавление строк пока i не станет больше или равно количеству строк таблицы(dTable.Rows.Count количество строк таблицы)
+                    for (int i = 0; i < dTable.Rows.Count; i++)
+                    {
+                        dataGridView1.Rows.Add(dTable.Rows[i].ItemArray);
+                    }
+                   
                 }
-            }
-            catch
-            {
+                catch
+                {
 
-            }
-            conbaza.Close();
+                }
+                conbaza.Close();         
         }
 
         private void change_service_Load(object sender, EventArgs e)
@@ -81,37 +89,24 @@ namespace Kursovaya_work
 
         private void change_service_FormClosed(object sender, FormClosedEventArgs e)
         {
-           
+            menu_director md = new menu_director();
+            md.Show();
+            this.Close();
         }
 
         private void button2_Click(object sender, EventArgs e)
         {
             string a = textBox3.Text;
-            try
+            if (textBox1.Text == "" || textBox2.Text == "" || textBox3.Text == "")
             {
-                conbaza.Open();
-                string commandStr = $"DELETE FROM Service WHERE ID_service = '{a}'";
-                MySqlDataAdapter adapter = new MySqlDataAdapter(commandStr, conbaza);
-                DataTable dTable = new DataTable();
-                adapter.Fill(dTable);                
+                MessageBox.Show("Введите данные");
             }
-            catch
+            else
             {
-
-            }
-            conbaza.Close();
-        
-        }
-
-        private void button3_Click(object sender, EventArgs e)
-        {
-            {
-                string a = textBox3.Text;
-                string b = textBox2.Text;
                 try
                 {
                     conbaza.Open();
-                    string commandStr = $"UPDATE Service SET cost_service = '{b}' WHERE ID_service = '{a}'";
+                    string commandStr = $"DELETE FROM Service WHERE ID_service = '{a}'";
                     MySqlDataAdapter adapter = new MySqlDataAdapter(commandStr, conbaza);
                     DataTable dTable = new DataTable();
                     adapter.Fill(dTable);
@@ -121,32 +116,64 @@ namespace Kursovaya_work
 
                 }
                 conbaza.Close();
+            }
+        
+        }
+
+        private void button3_Click(object sender, EventArgs e)
+        {
+            {
+                string a = textBox3.Text;
+                string b = textBox2.Text;
+                if (textBox1.Text == "" || textBox2.Text == "" || textBox3.Text == "")
+                {
+                    MessageBox.Show("Введите данные");
+                }
+                else
+                {
+                    try
+                    {
+                        conbaza.Open();
+                        //Обновить значения cost_service и ID_service в таблице Service
+                        string commandStr = $"UPDATE Service SET cost_service = '{b}' WHERE ID_service = '{a}'";
+                        MySqlDataAdapter adapter = new MySqlDataAdapter(commandStr, conbaza);
+                        DataTable dTable = new DataTable();
+                        adapter.Fill(dTable);
+                    }
+                    catch
+                    {
+
+                    }
+                    conbaza.Close();
+                }
 
             }
         }
 
         private void button4_Click(object sender, EventArgs e)
         {
-            try
-            {
-                conbaza.Open();
-                //Запрос к базе данных(взять ID_service,service,cose_service из таблицы service)
-                string commandStr = "SELECT ID_service,service,cost_service FROM Service";
-                MySqlDataAdapter adapter = new MySqlDataAdapter(commandStr, conbaza);
-                DataTable dTable = new DataTable();
-                adapter.Fill(dTable);
-                dataGridView1.Rows.Clear();
-                //Добавление строк пока i не станет больше или равно количеству строк таблицы(dTable.Rows.Count количество строк таблицы)
-                for (int i = 0; i < dTable.Rows.Count; i++)
+          
+                try
                 {
-                    dataGridView1.Rows.Add(dTable.Rows[i].ItemArray);
+                    conbaza.Open();
+                    //Запрос к базе данных(взять ID_service,service,cose_service из таблицы service)
+                    string commandStr = "SELECT ID_service,service,cost_service FROM Service";
+                    MySqlDataAdapter adapter = new MySqlDataAdapter(commandStr, conbaza);
+                    DataTable dTable = new DataTable();
+                    adapter.Fill(dTable);
+                    dataGridView1.Rows.Clear();
+                    //Добавление строк пока i не станет больше или равно количеству строк таблицы(dTable.Rows.Count количество строк таблицы)
+                    for (int i = 0; i < dTable.Rows.Count; i++)
+                    {
+                        dataGridView1.Rows.Add(dTable.Rows[i].ItemArray);
+                    }
                 }
-            }
-            catch
-            {
+                catch
+                {
 
-            }
-            conbaza.Close();
+                }
+                conbaza.Close();
+            
         }
 
         private void button6_Click(object sender, EventArgs e)
